+++ src/components/Headline.tsx
@@ -1,6 +1,4 @@
-import * as React from 'react';
-import Typography from '@mui/material/Typography';
-import Box from '@mui/material/Box';
+import { Typography, Box } from '@mui/material';
 import TagChip from './TagChip';
 
 interface HeadlineProps {
